[shadows] Fix shadows with small blur radius
There's an assertion in calculate_gaussian_kernel() to avoid a division
by zero - due to an unnecessary cast from float to int this assertion
is triggered incorrectly for small (but non-zero) values, e.g. a blur
radius of 1px.

+++ src/st/st-theme-node-drawing.c
@@ -50,7 +50,7 @@ calculate_gaussian_kernel (gdouble   sigma,
   gdouble exp_divisor;
   gint half, i;
 
-  g_return_val_if_fail ((int) sigma > 0, NULL);
+  g_return_val_if_fail (sigma > 0, NULL);
 
   half = n_values / 2;
